+++ vscode-extension/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.1.1
 
-Update Az module target version from 4.6.1 to 4.7.0
+Update Az module target version from 4.6.1 to 4.8.0
 
 ## 0.1.0