+++ vscode-extension/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Azure PowerShell Extension Release History
 
+## 0.2.0
+
+Code refactoring and user experience optimization:
+- Remove the language server
+- Change the activation event to openning the Powershell file
+- Start the powerShell process only once
+- Fixed bug when opening multiple files quickly
+
+
 ## 0.1.0
 
 First preview release with the following features: